requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
certifi==2020.6.20
chardet==3.0.4
Django==3.0.3
django-ratelimit==3.0.1
djangorestframework==3.11.0
idna==2.10
psycopg2-binary==2.8.5
python-memcached==1.59
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
